+++ build.sh
@@ -38,7 +38,7 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 declare -A extra_tags
 extra_tags=( ["nompi"]="-t ${repository}/software:latest"
              ["greatlakes"]=""
-             ["comet"]=""
+             ["expanse"]=""
              ["bridges2"]=""
              ["stampede2"]=""
 )