+++ docker/expanse/test/job-gpu.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+#SBATCH --job-name="test-gpu"
+#SBATCH --partition=gpu-shared
+#SBATCH --nodes=1
+#SBATCH --ntasks-per-node=1
+#SBATCH --gpus=1
+#SBATCH -t 0:10:00
+
+rm -f test-results-gpu.out
+
+singularity exec --nv software.sif python3 serial-gpu.py